+++ src/comments.ts
@@ -307,6 +307,40 @@ function handleBinaryishExpressionRightChildTrailingComment(
 }
 
 /**
+ * In a if-else block, if there is an end of line comment for the inner statement, we want to
+ * attach it to that statement on the right. This doesn't work by default if we force adding curly braces around the if/else block.
+ * So this method does it instead.
+ */
+function handleIfElseBlockRightTrailingComment(
+  comment: AnnotatedComment,
+  options: ParserOptions,
+) {
+  const { precedingNode } = comment;
+  if (
+    !options.apexForceCurly ||
+    comment.placement !== "endOfLine" ||
+    !precedingNode
+  ) {
+    return false;
+  }
+  if (
+    precedingNode["@class"] === "apex.jorje.data.ast.IfBlock" ||
+    precedingNode["@class"] === "apex.jorje.data.ast.ElseBlock"
+  ) {
+    addTrailingComment(precedingNode.stmnt, comment);
+    return true;
+  }
+  if (precedingNode["@class"] === "apex.jorje.data.ast.Stmnt$IfElseBlock") {
+    const commentLeftStatement =
+      precedingNode.elseBlock?.value?.stmnt ??
+      precedingNode.ifBlocks[precedingNode.ifBlocks.length - 1].stmnt;
+    addTrailingComment(commentLeftStatement, comment);
+    return true;
+  }
+  return false;
+}
+
+/**
  * Turn the leading comment in a long method or variable chain into the preceding
  * comment of a previous node. Without doing that, we have an awkward position
  * for the . character like so:
@@ -397,6 +431,7 @@ export function handleOwnLineComment(
  *
  * @param comment The comment node.
  * @param sourceCode The entire source code.
+ * @param options The prettier options.
  * @returns {boolean} Whether we have manually attached this comment to some AST
  * node. If `true` is returned, Prettier will no longer try to attach this
  * comment based on its internal heuristic.
@@ -404,10 +439,12 @@ export function handleOwnLineComment(
 export function handleEndOfLineComment(
   comment: AnnotatedComment,
   sourceCode: string,
+  options: ParserOptions,
 ): boolean {
   return (
     handleDanglingComment(comment) ||
     handleBinaryishExpressionRightChildTrailingComment(comment) ||
+    handleIfElseBlockRightTrailingComment(comment, options) ||
     handleBlockStatementLeadingComment(comment) ||
     handleWhereExpression(comment, sourceCode) ||
     handleModifierPrettierIgnoreComment(comment) ||