+++ src/printer.ts
@@ -197,6 +197,25 @@ function handleExpandedBlockStatement(
     : statement;
 }
 
+function handleBracesAroundStatement(
+  statement: Doc,
+  statementType: Doc,
+  options: ParserOptions,
+): Doc {
+  const parts: Doc[] = [];
+  if (statementType === APEX_TYPES.BLOCK_STATEMENT) {
+    parts.push(" ");
+    pushIfExist(parts, statement);
+  } else if (options.apexForceCurly) {
+    parts.push(" ");
+    parts.push(group(indent(["{", hardline, statement])));
+    parts.push([hardline, "}"]);
+  } else {
+    pushIfExist(parts, group(indent(concat([hardline, statement]))));
+  }
+  return parts;
+}
+
 function escapeString(text: string): string {
   // Code from https://stackoverflow.com/a/11716317/477761
   return text
@@ -1816,8 +1835,14 @@ function handleNewExpression(path: AstPath, print: printFn): Doc {
   return concat(parts);
 }
 
-function handleIfElseBlock(path: AstPath, print: printFn): Doc {
+// TODO: Add the same behavior for other blocks: for/while/do-while/try-catch-finally
+function handleIfElseBlock(
+  path: AstPath,
+  print: printFn,
+  options: ParserOptions,
+): Doc {
   const node = path.getValue();
+  const forceCurly = options.apexForceCurly;
   const parts: Doc[] = [];
   const ifBlockDocs: Doc[] = path.map(print, "ifBlocks");
   const elseBlockDoc: Doc = path.call(print, "elseBlock", "value");
@@ -1851,7 +1876,8 @@ function handleIfElseBlock(path: AstPath, print: printFn): Doc {
         !ifBlockContainsBlockStatement[index - 1] ||
         ifBlockContainsLeadingOwnLineComments[index] ||
         ifBlockContainsTrailingComments[index - 1];
-      if (shouldAddHardLineBeforeElseIf) {
+
+      if (shouldAddHardLineBeforeElseIf && !forceCurly) {
         parts.push(hardline);
       } else {
         parts.push(" ");
@@ -1861,7 +1887,7 @@ function handleIfElseBlock(path: AstPath, print: printFn): Doc {
     // We also need to handle the last if block, since it might need to add
     // either a space or a hardline before the else block
     if (index === ifBlockDocs.length - 1 && elseBlockDoc) {
-      if (ifBlockContainsBlockStatement[index]) {
+      if (ifBlockContainsBlockStatement[index] || forceCurly) {
         parts.push(" ");
       } else {
         parts.push(hardline);
@@ -1883,6 +1909,7 @@ function handleIfElseBlock(path: AstPath, print: printFn): Doc {
       !lastIfBlockHardLineInserted &&
       (elseBlockContainsLeadingOwnLineComments ||
         lastIfBlockContainsTrailingComments);
+
     if (shouldAddHardLineBeforeElse) {
       parts.push(hardline);
     }
@@ -1891,7 +1918,11 @@ function handleIfElseBlock(path: AstPath, print: printFn): Doc {
   return groupConcat(parts);
 }
 
-function handleIfBlock(path: AstPath, print: printFn): Doc {
+function handleIfBlock(
+  path: AstPath,
+  print: printFn,
+  options: ParserOptions,
+): Doc {
   const node: EnrichedIfBlock = path.getNode();
   const statementType: Doc = path.call(print, "stmnt", "@class");
   const statementDoc: Doc = path.call(print, "stmnt");
@@ -1912,29 +1943,23 @@ function handleIfBlock(path: AstPath, print: printFn): Doc {
   conditionParts.push(")");
   parts.push(groupIndentConcat(conditionParts));
   // Body block
-  if (statementType === APEX_TYPES.BLOCK_STATEMENT) {
-    parts.push(" ");
-    pushIfExist(parts, statementDoc);
-  } else {
-    pushIfExist(parts, group(indent(concat([hardline, statementDoc]))));
-  }
-  return concat(parts);
+  parts.push(handleBracesAroundStatement(statementDoc, statementType, options));
+  return parts;
 }
 
-function handleElseBlock(path: AstPath, print: printFn): Doc {
+function handleElseBlock(
+  path: AstPath,
+  print: printFn,
+  options: ParserOptions,
+): Doc {
   const statementType: Doc = path.call(print, "stmnt", "@class");
   const statementDoc: Doc = path.call(print, "stmnt");
 
   const parts: Doc[] = [];
   parts.push("else");
   // Body block
-  if (statementType === APEX_TYPES.BLOCK_STATEMENT) {
-    parts.push(" ");
-    pushIfExist(parts, statementDoc);
-  } else {
-    pushIfExist(parts, group(indent(concat([hardline, statementDoc]))));
-  }
-  return concat(parts);
+  parts.push(handleBracesAroundStatement(statementDoc, statementType, options));
+  return parts;
 }
 
 function handleTernaryExpression(path: AstPath, print: printFn): Doc {