+++ tests/conditional_force_curly/jsfmt.spec.ts
@@ -1,0 +1,4 @@
+runSpec(__dirname, ["apex"], {
+  apexForceCurly: true,
+  astCompareDisabled: true,
+});